requirements.txt
fastapi
uvicorn[standard]
pillow
pandas
python-dotenv
easyocr
langchain_google_genai
